requirements.txt
asgiref
Django
djangorestframework
djangorestframework_simplejwt
PyJWT
sqlparse
tzdata
